Add a NAT adapter for connecting to the console
In a previous commit I had to remove all of the NAT adapters because
they seemed to cuase the microkernel download on the PXE boot box
freeze.  I want users to be able to connect to the PE console
so I added the NAT adapter on the puppet-master back in and
everything appears to be working.