Update dhcp config to ensure it works if razor server isn't up
I updated the razor_dnsmasq module to retry pulling down the
microkernal from the razor server if for some reason the wget
fails.  This seems to be an issue for some users if the razor
server takes a while to install and the dhcp server comes up
quickly.

I also updated the PE version to 3.2.3