Add domains to the host names
Prior to this commit, if you connected to a network that would
assign your vm a domain via DHCP then your install of PE would
fail.  This is because the installer would look for
puppet-master.yourdomain.com and it wouldn't be in /etc/hosts so
it'd all fail.

After this commit, the VM is defined a hostname with an arbitrary
domain and this resolves the issue for some reason I can't remember.
I think it's because the hosts vagrant plugin creates an /etc/hosts
entry for both the short name and the fqdn so if anything refers to
the fqdn then it is resolvable.