(maint) Update sample PXE image and docs, make servers homogeneous
This includes updating the README to address needed changes in the
supplied example VM image, that image and changes to the vms.yaml file to
have razor-server come up in VirtualBox instead of VMWare.

Changes to the example VM include changing the base memory to 2048Mb and
specifying boot order to have network first such that rebooting and reinstall
work properly.